pyproject.toml
[tool.poetry]
name = "nlp-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sklearn = "^0.0"
matplotlib = "^3.4.1"
seaborn = "^0.11.1"
panda = "^0.3.1"
jupyter = "^1.0.0"
ipykernel = "^5.5.3"
tensorflow-datasets = "^4.2.0"
numpy = "<1.20.0"
wheel = "^0.36.2"
grpcio = "^1.32.0"
nltk = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
